pyproject.toml
[tool.poetry]
name = "pubmedfetcher"
version = "0.1.0"
description = "PubMed Paper Fetcher"
authors = ["Aniket Shewale <aniketshewale26@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
biopython = "^1.79"
typer = "^0.9.0"
pandas = "^2.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"

[tool.poetry.scripts]
pubmedfetcher = "src.pubmedfetcher.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


